requirements.txt
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.0.4
Flask==0.12.5
Flask-APIBlueprint==1.0.0
idna==3.3
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.29
protobuf==3.19.4
PyJWT==2.4.0
python-dotenv==0.20.0
python-logstash==0.4.8
redis==4.3.3
requests==2.27.1
six==1.16.0
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==0.16.1
zipp==3.6.0
